Form, Header & Dashboard
For better maintainability and scalability I'm going to be breaking the code into small components to re-utilize them as the development process moves forward.

Such is the case of Input.tsx inside components/Form, this way we can better understand the code and keep it organized where each component deals with their respective responsibilities withing the rules of business of our app.

Header:
Created a Header component, with a different input, henceforth I'm not using the Input we set as default within our app.

Dashboard:
Created the first components and styles for the Dashboard page.