Sidebar - Part 1
Started the Sidebar component with it's first styling , just to have a visual reference of what it should look like during the development.

This Sidebar will be displayed in most of the pages of the app, and further down the development process I will be splitting this code into smaller components for better readability.