Sidebar - Part 2
As previous disclosed I was going to separate  the code withing the components to further increase the organization and scalability of the project.

Taking a look at the new structure of the component SideBar in contrast to the SideBar.tsx file, it's realizable that improves significantly the readability and scalability of the code.

Sidebar:
+NavSection
+NavLink

 NavSection:
Created a component for each section of the SideBar component, typing on NavSection the title that can receive different data.

 NavLink:
Created a component for the links typing the title and the icon, since the link structure require an icon and a title. This way we can easily pass the data we need in the SideBar/index.tsx