Pagination Component  - Part 2
Following the same process of the SideBar component now on the Pagination component.

Pagination:
+PaginationItem

 PaginationItem:
Since we have two types of buttons in the PaginationComponent I typed a conditional variant to as isCurrent? and a number type for the number of pages. I used the conditional type to display both variations of the button. It's not functional for now only for visual reference.