Formatting the Date
Now I can receive the data from the generated users and format the date of the createdAt field to more readable field. Also learned to fix the data in a useQuery Hook before the component receive any information. This way the developer can take measures to fix problematic data before it's sent to the components to consume it, which is very neat.