Managing Server-Side States with JQuery
Using the isFetching method from JQuery we can signalize the user that the app is requesting data, portrait by the spinner added next to 'Users' title in the users interface. This way we setup two different spinners, one for the first fetch of the data, and the other when the data is being re-fetched even when the data is cached.