Custom Hook useUser
Separated the logic from the fetch data to a custom hook component and there i split the code into two functions, one for the custom hook and other to deal with the fetch logic. Also typed the return of this function as <User> this way anytime we call this logic, we will be able to see clearly the props from the data we're dealing with.