Pagination Component - Part 3
Started to develop the logic behind the pagination component. Since in the visual guide the Pagination should show the first page, the previous page, the current page, the next page and the last page, I've implemented the logic for that too.

Also reduced the size of the button which is active to fit the size of the others, the visual queue is enough to signalize to the user the current page.